requirements.txt
django
django-autocomplete-light
django-admin-rangefilter